Implements snicker transaction tweaking and persists scanned candidates.
The tweaking is done with ECDH on the proposer's ephemeral and the receiver's output pubkey. Taproot only of course. The proposals are encrypted using AES-GCM with the shared secret as key. The format of proposals is (ephemeral_pubkey, tag), (encrypted_proposal) where the tag is a fixed derivation of the DH shared secret and so easy for receivers to filter without giving data to external parties.

This is currently mostly a skeleton as e.g. transaction construction has not been done, nor any mechanism for proposal broadcast.