Adds automation functions
Configurable mode of automation, allowing
background scanning for proposals (currently
proposals are manually shared as files, into a specific
proposals directory; in future the network layer
can populate this location).
The database schema is simplified; only the decrypted
proposals are stored in it with a status field to remember
if they have been processed/accepted/broadcast.
Automation is configurable and with one specifically important
feature: prioritizing SNICKER-patterned p2tr transactions for
proposal (creating chains because of higher likelihood of
acceptance).
The CLI implements automation but the GUI does not, yet.