Automation in GUI; subscribe not poll
This has a lot of fixes, perhaps most importantly removing
any polling which can cause CPU churn. We subscribe to Kyoto events
and wrap that in the wallet node so the API consumers can subscribe to us.
The same automation that was on CLI now also works in the GUI, though
only the Receiver side has been tested.
The wallet's utxos are now properly wrapped such that the same actions,
including accepting proposals for snicker coinjoins, can occur for existing
snicker utxos just as for normal utxos.
Finally the test harness was updated though one specific test is still bugged.