Big refactor to maintain partial utxo state
Realizing that a "normal" light client model just doesn't work,
realistically: that e.g. a receiver cannot validate the proposals
from proposers because it cant check their inputs are real (notice
this is a function of trying to do *coinjoin* and not just normal
wallet operations, with a light client), I realised that the best
model to deal with this was to maintain a small, filtered view of
the full utxo set: limited to (birth date-1000 blocks, limited to
p2tr, and limited to >5k sats), we can ensure we're correctly both
our own utxos and those of proposals sent to us.
This is obviously a very large change and involves adding another
database file (partial_utxo_set.db) to persistence.
There are also some slight updates to the nostr networking feature,
and a preliminary draft of a GUI status bar.